+++ lib/l10n/app_localizations_cs.dart
@@ -2287,6 +2287,9 @@ class AppLocalizationsCs extends AppLocalizations {
   String get backfillStatsBackfilled => 'Doplněno';
 
   @override
+  String get backfillStatsBurned => 'Anulováno';
+
+  @override
   String get backfillStatsDeleted => 'Smazáno';
 
   @override